@@ -1,6 +1,6 @@
 {
   "name": "crowi",
-  "version": "1.6.0",
+  "version": "1.6.1",
   "description": "The simple & powerful Wiki",
   "tags": [
     "wiki",
@@ -31,10 +31,13 @@
     "async": "~1.5.0",
     "aws-sdk": "~2.2.26",
     "axios": "0.15.x",
-    "babel-core": "~6.7.6",
-    "babel-loader": "~6.2.4",
-    "babel-preset-es2015": "~6.6.0",
-    "babel-preset-react": "~6.5.0",
+    "babel-cli": "~6.24.1",
+    "babel-core": "~6.24.1",
+    "babel-loader": "~7.0.0",
+    "babel-polyfill": "~6.23.0",
+    "babel-preset-env": "~1.4.0",
+    "babel-preset-es2015": "~6.24.1",
+    "babel-preset-react": "~6.24.1",
     "basic-auth-connect": "~1.0.0",
     "body-parser": "~1.14.1",
     "bootstrap-sass": "~3.3.6",
@@ -54,7 +57,7 @@
     "elasticsearch": "^12.1.3",
     "emojify.js": "^1.1.0",
     "errorhandler": "~1.3.4",
-    "express": "~4.14.0",
+    "express": "~4.15.2",
     "express-form": "~0.12.0",
     "express-session": "~1.14.0",
     "font-awesome": "~4.7.0",
@@ -88,8 +91,9 @@
     "multer": "~1.2.1",
     "nodemailer": "~2.7.0",
     "nodemailer-ses-transport": "~1.5.0",
-    "react": "~15.0.1",
-    "react-dom": "~15.0.1",
+    "react": "~15.5.0",
+    "react-bootstrap": "~0.30.10",
+    "react-dom": "~15.5.0",
     "redis": "~2.6.5",
     "reveal.js": "~3.2.0",
     "socket.io": "~1.3.0",
@@ -97,9 +101,9 @@
     "sprintf": "~0.1.5",
     "swig": "~1.4.0",
     "vinyl-source-stream": "~1.1.0",
-    "webpack": "~1.13.0",
-    "webpack-manifest-plugin": "~1.0.1",
-    "webpack-stream": "~3.1.0"
+    "webpack": "~2.2.0",
+    "webpack-manifest-plugin": "~1.1.0",
+    "webpack-stream": "~3.2.0"
   },
   "devDependencies": {
     "chai": "~1.10.0",
