@@ -34,7 +34,7 @@ module.exports = function(crowi, app) {
       fallbackLng: [User.LANG_EN_US],
       whitelist: Object.keys(User.getLanguageLabels()).map((k) => User[k]),
       backend: {
-        loadPath: 'locales/{{lng}}/translation.json'
+        loadPath: crowi.localeDir + '{{lng}}/translation.json'
       },
       detection: {
         order: ['userSettingDetector', 'header', 'navigator'],
@@ -71,6 +71,7 @@ module.exports = function(crowi, app) {
         language:   User.getLanguageLabels(),
         registrationMode: Config.getRegistrationModeLabels(),
     };
+    res.locals.local_config = Config.getLocalconfig(config); // config for browser context
 
     next();
   });
