@@ -353,6 +353,8 @@ module.exports = function(crowi) {
       path = '/' + path;
     }
 
+    path = path.replace(/\/\s+?/g, '/').replace(/\s+\//g, '/');
+
     return path;
   };
 
@@ -394,6 +396,8 @@ module.exports = function(crowi) {
       /^\/_r\/.*/,
       /^\/user\/[^\/]+\/(bookmarks|comments|activities|pages|recent-create|recent-edit)/, // reserved
       /^\/?https?:\/\/.+$/, // avoid miss in renaming
+      /\/{2,}/,             // avoid miss in renaming
+      /\s+\/\s+/,           // avoid miss in renaming
       /.+\/edit$/,
       /.+\.md$/,
       /^\/(installer|register|login|logout|admin|me|files|trash|paste|comments)(\/.*|$)/,
@@ -546,12 +550,34 @@ module.exports = function(crowi) {
     });
   };
 
+  pageSchema.statics.findPageByRedirectTo = function(path) {
+    var Page = this;
+
+    return new Promise(function(resolve, reject) {
+      Page.findOne({redirectTo: path}, function(err, pageData) {
+        if (err || pageData === null) {
+          return reject(err);
+        }
+
+        return resolve(pageData);
+      });
+    });
+  };
+
   pageSchema.statics.findListByCreator = function(user, option, currentUser) {
     var Page = this;
     var User = crowi.model('User');
     var limit = option.limit || 50;
     var offset = option.offset || 0;
-    var conditions = { creator: user._id, redirectTo: null };
+    var conditions = {
+      creator: user._id,
+      redirectTo: null,
+      $or: [
+        {status: null},
+        {status: STATUS_PUBLISHED},
+      ],
+    };
+
     if (!user.equals(currentUser._id)) {
       conditions.grant = GRANT_PUBLIC;
     }
@@ -908,8 +934,10 @@ module.exports = function(crowi) {
       }).then(function(done) {
         return Page.removePageById(pageId);
       }).then(function(done) {
+        return Page.removeRedirectOriginPageByPath(pageData.path);
+      }).then(function(done) {
         pageEvent.emit('delete', pageData, user); // update as renamed page
-        resolve();
+        resolve(pageData);
       }).catch(reject);
     });
   };
@@ -927,6 +955,43 @@ module.exports = function(crowi) {
         resolve(done);
       });
     });
+  };
+
+  pageSchema.statics.removePageByPath = function(pagePath) {
+    var Page = this;
+
+    return Page.findPageByPath(redirectPath)
+      .then(function(pageData) {
+        return Page.removePageById(pageData.id);
+      });
+  };
+
+  /**
+   * remove the page that is redirecting to specified `pagePath` recursively
+   *  ex: when
+   *    '/page1' redirects to '/page2' and
+   *    '/page2' redirects to '/page3'
+   *    and given '/page3',
+   *    '/page1' and '/page2' will be removed
+   *
+   * @param {string} pagePath
+   */
+  pageSchema.statics.removeRedirectOriginPageByPath = function(pagePath) {
+    var Page = this;
+
+    return Page.findPageByRedirectTo(pagePath)
+      .then((redirectOriginPageData) => {
+        // remove
+        return Page.removePageById(redirectOriginPageData.id)
+          // remove recursive
+          .then(() => {
+            return Page.removeRedirectOriginPageByPath(redirectOriginPageData.path)
+          });
+      })
+      .catch((err) => {
+        // do nothing if origin page doesn't exist
+        return Promise.resolve();
+      })
   };
 
   pageSchema.statics.rename = function(pageData, newPagePath, user, options) {
